+++ R/redcap-read-oneshot.R
@@ -253,7 +253,7 @@ redcap_read_oneshot <- function(
   checkmate::assert_list(     config_options            , any.missing=TRUE ,        null.ok=TRUE)
   # checkmate::assert_character(encode_httr               , any.missing=FALSE, len=1, null.ok = FALSE)
 
-  validate_field_names(fields, stop_on_error = TRUE)
+  assert_field_names(fields)
 
   pseudofields <- c(
     "redcap_event_name",