+++ inst/test-data/specific-redcapr/read-batch-simple/specify-fields-without-record-id.R
@@ -1,0 +1,6 @@
+structure(list(record_id = c(1, 2, 3, 4, 5), name_first = c("Nutmeg", 
+"Tumtum", "Marcus", "Trudy", "John Lee"), address = c("14 Rose Cottage St.\nKenning UK, 323232", 
+"14 Rose Cottage Blvd.\nKenning UK 34243", "243 Hill St.\nGuthrie OK 73402", 
+"342 Elm\nDuncanville TX, 75116", "Hotel Suite\nNew Orleans LA, 70115"
+), interpreter_needed = c(0, 0, 1, NA, 0)), row.names = c(NA, 
+-5L), class = c("spec_tbl_df", "tbl_df", "tbl", "data.frame"))