+++ inst/test-data/specific-redcapr/read-oneshot/specify-fields-without-record-id.R
@@ -1,0 +1,11 @@
+structure(list(name_first = c("Nutmeg", "Tumtum", "Marcus", "Trudy", 
+"John Lee"), address = c("14 Rose Cottage St.\nKenning UK, 323232", 
+"14 Rose Cottage Blvd.\nKenning UK 34243", "243 Hill St.\nGuthrie OK 73402", 
+"342 Elm\nDuncanville TX, 75116", "Hotel Suite\nNew Orleans LA, 70115"
+), interpreter_needed = c(0, 0, 1, NA, 0)), row.names = c(NA, 
+-5L), spec = structure(list(cols = list(name_first = structure(list(), class = c("collector_character", 
+"collector")), address = structure(list(), class = c("collector_character", 
+"collector")), interpreter_needed = structure(list(), class = c("collector_double", 
+"collector"))), default = structure(list(), class = c("collector_guess", 
+"collector")), delim = ","), class = "col_spec"), class = c("spec_tbl_df", 
+"tbl_df", "tbl", "data.frame"))