+++ utility/refresh.R
@@ -40,7 +40,7 @@ devtools::run_examples(); #dev.off() #This overwrites the NAMESPACE file too
 # pkgload::load_all()
 test_results_checked <- devtools::test()
 test_results_checked <- devtools::test(filter = "read-batch-survey")
-test_results_checked <- devtools::test(filter = "^metadata-coltypes")
+test_results_checked <- devtools::test(filter = "^validate")
 withr::local_envvar(ONLYREADTESTS = "true")
 test_results_checked <- devtools::test(filter = "write-batch")